Workaround ABI differences between stdc++ and libc++
In Qt6, more standard library types are exposed in the public API. However,
the ABI of some of those types is different between the two standard library
implementations - an example for all is std::exception_ptr, which is defined
as std::__exception_ptr::exception_ptr in libstdc++. This causes problems
when Qt is built with one standard library but QCoro is being built using
another one - typically Qt is built against libstdc++ but we use libc++ when
building QCoro with Clang. This causes linking error because Clang/libc++
looks for std::exception_ptr symbol, but Qt has std::__exception_ptr::exception_ptr.

So far I've seen this problem manifest only in QUnhandledException, which
unfortunately is used in QtConcurrent, so we disable QFuture support in
QCoro when we detect this mismatch at configure time.