Implement operations provided by qCoro() as regular coroutines
This means that every operation provided by the QCoro wrapper
classes now returns a Task<T>, so it's possible to use .then()
continuations with any operations provided by QCoro, e.g.

qCoro(socket).waitForConnected(then([](bool ok) { ... });

Previously qCoro(socket).waitForConnected() would return an
instance of QCoroAbstractSocket::WaitForConectedOperation,
which is an implementation detail of QCoro, which makes it
hard to combine it with functions (not coroutines!) that
return Task<T> and makes it much harder to support .then()
continuations everywhere.