Fix ODR violation when LTO is enabled
There are two classes called SocketReadySignalHelper in two
anonymous namespaces in two distinct TUs. Since they are both
used as arguments to a coroutine, the compiler has generated for
each invocation of the coroutine a stackframe struct, which
contained SocketReadySignalHelper in its name. However, the frame
structs are not stored in an anonymous namespace, and due to the
classes both having the same name, we ended up with two distict
classes with the same name in two different TUs. Because they are
not in an anonymous namespace, the link time optimizer tries to
solve the duplication and runs into an ODR violation.

To workaround this, what I believe is a GCC bug, this change renames
the structures to each have a unique name.

The name of the generated frame struct was
_Z5qCoroIN12_GLOBAL__N_123SocketReadySignalHelperEMN5QCoro6detail16WaitSignalHelperEFvbEENS2_4TaskINS3_11QCoroSignalIT_T0_E11result_typeEEEPS9_OSA_NSt6chrono8durationIlSt5ratioILl1ELl1000EEEE.Frame