Make sure the test server waits for incoming connection
The socket state is local, so it's possible to report 'connected',
resume to coroutine finish it and destroy the socket before the
server has time to actually get the incoming connection. This
results in the test server being stopped by the test before it
actually starts waiting for incoming connection, causing the
server to incorrectly report the test as time out.

The fix is for the test to wait at the end for the server to
confirm it has gotten the incoming connection.