Explicitly enable exception support for clang-cl (fixes #90)
MSVC requires that exceptions are explicitly enabled using the `\EH`
flag, so clang-cl copies that behavior. When CMake targets the
Windows-MSVC platform (that is when using "Visual Studio" generator),
CMake takes care of adding `\EHsc` to `CXXFLAGS`. However when the
clang-cl compiler is specified manually (`CMAKE_CXX_COMPILER`) and
another generator, like Ninja is used, then CMake ends up targeting
the Windows-Clang-CXX platform, which does not enable exceptions,
so we have to do that ourselves. To keep the code simple, we do not
try to detect the target platform or such, we simply append the `\EHsc`
flag when we detect clang-cl.