Introduce AbortOnException option to QCoro::Task
Adds an additional optinal template argument to QCoro::Task,
which may specify an AbortOnException option. When the option
is set on a coroutine and the coroutine throws an exception,
QCoro will abort the program, UNLESS the coroutine is being
co_awaited, which is the only way to propagate the exception
from a coroutine.