Add QCoro::Task<T>::then overload with context object
This solves the same problem as the 4-arg QObject::connect function.

Without listening for the context object being destroyed, a callback
that captures the a pointer from the context could access deleted
memory, if the pointed-to object is deleted while awaiting.

With this change, we will just not invoke the callback if the context
object was deleted in the meantime.