Fix QCoroFuture::waitForFinished triggerring twice on cancelation
While not explicitly documented, when a QFuture is canceled
it will enter both cancelled and finished state (and emit both
signals). Since we were listening to both signals, a canceled
QFuture would cause QCoroFuture to resume the awaiter once on
the canceled signal and then again when the finished() signal
is emitted, possibly leading to a crash when the awaiter has
already been destroyed after the first resume.

More details in discussion #108.