tests: Construct a dummy QWebSocket before starting tests
On Windows, the first QWebSocket instantiation takes a long time,
likely due to OpenSSL being dynamically loaded and resolved. On the
Github CI, this takes so long, that the testcase times out. As
a workaround, we construct a dummy QWebSocket in test's constructor
so that the initialization happens before the actual test starts.

Resolves #116.