Handle when a task is co_awaited by multiple awaiters
The old approach caused that the last awaiter to co_await a coroutine
function has overwritten the reference to the previous (so that any
previous awaiter would not be resumed when the coroutine function has
finished) but at the same time the `await_suspend()` would return `false`,
so that the awaiter would get resume again immediatelly. However, since
the last awaiter has also been registered as the only awaiting coroutine,
once the awaited coroutine has finished, the awaiter was resumed for the
second time, even if not suspended, causing crashes or memory leaks.

This change makes `TaskAwaiterBase::await_suspend()` a void function,
and stores awaiters as a list, so that they can all get resumed when
the coroutine finishes.