Fix crash in QCoroSignal when signal is received after destruction
When there's a signal emission pending while QCoroSignal is destroyed,
it will still get delivered, but the this pointer is already destroyed,
which leads to a crash.

This is fixed by using a dedicated QObject as the signal receiver when
setting up the connection which is destroyed with the QCoroSignal.

Fixes #167.