Prevent coroutine from being stuck if sender is deleted
It is possible for a sender QObject to be deleted before it gets a
chance to emit a signal that we are co_await'ing on. In such a case, the
coroutine will never be resumed. A side-effect of this is that all
variables that are local to that coroutine will never get deallocated
and will stay in limbo forever.