Use refcount to track when to destroy coroutine
Instead of a single bool that tells us whether we can destroy the
coroutine safely, we now use a ref-count and destroy the coroutine
when the count drops to 0. Generally, a coroutine has ref count
value of 2 - one is held by the coroutine itself (ensuring it's not
deleted until it finishes) and one is held by QCoro::Task, ensuring
we can access the coroutine state as long as the corresponding Task
exists.

This also opens doors to future enhancements, like SharedTasks that
can be safely co_awaited by multiple awaiters.